Defined a function for setting up org-mode
- org-indent-mode: When active, indent text according to outline
structure.

- the commented out part makes everything in a variable width font
which needs extra configuration to make sure that text that is meant
to be displayed in fixed width font gets displayed in that font

- visual-line-mode: Toggle visual line based editing (Visual Line
mode) in the current buffer. When Visual Line mode is enabled,
`word-wrap' is turned on in this buffer, and simple editing commands
are redefined to act on visual lines, not logical lines.